requirements.txt
schedule
pycryptodome
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
ttkbootstrap
pycryptodome
